[Unification] Consolidate SiLU for CLIP and GPT
[ghstack-poisoned]
