+++ torchmultimodal/models/gpt.py
@@ -10,6 +10,7 @@ from typing import Any, Callable, Dict, NamedTuple, Optional, Tuple, Union
 import torch
 from torch import nn, Tensor
 
+from torchmultimodal.modules.layers.activation import SiLU
 from torchmultimodal.modules.layers.attention import MultiHeadAttention, SelfAttention
 from torchmultimodal.modules.layers.mlp import MLP
 from torchmultimodal.utils.attention import get_extended_attention_mask
@@ -506,22 +507,6 @@ class TransformerDecoder(nn.Module):
             attention_weights=all_attentions,
             past_key_values=all_past_key_values,
         )
-
-
-class SiLU(nn.Module):
-    r"""Sigmoid Linear Unit
-
-    .. math:: \text{SiLU}(x) = x * \sigma(1.702 * x)
-
-    where :math:`\sigma(x)` is the cumulative distribution function for Logistic Distribution.
-
-    Approximation of the exact GeLU for greater forward speed. Note that this is different from
-    ``torch.nn.SiLU`` by the coefficient ``1.702`` from the paper:
-    `"Gaussian error linear units"<https://arxiv.org/pdf/1606.08415.pdf>`_.
-    """
-
-    def forward(self, x: Tensor) -> Tensor:
-        return torch.sigmoid(1.702 * x) * x
 
 
 class TransformerDecoderLayer(nn.Module):
